% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographicProjectionApp.R
\name{demographicProjectionApp}
\alias{demographicProjectionApp}
\title{Run the Bayesian caribou demographic projection app}
\usage{
demographicProjectionApp(n = 1000)
}
\value{
The app will launch in the browser
}
\description{
The app will launch in the browser.
}
\examples{
demographicProjectionApp()

}
